[macos] prepare repository for ARM64 arch (#7396)

+++ images/macos/provision/core/azcopy.sh
@@ -1,15 +1,21 @@
 #!/bin/bash -e -o pipefail
 source ~/utils/utils.sh
+arch=$(get_arch)
 
-AZCOPY_DOWNLOAD_URL="https://aka.ms/downloadazcopy-v10-mac"
+# Check MacOS architecture and if ARM install using brew
+if [ $arch == "arm64" ]; then
+    brew install azcopy
+else
+    AZCOPY_DOWNLOAD_URL="https://aka.ms/downloadazcopy-v10-mac"
 
-download_with_retries $AZCOPY_DOWNLOAD_URL "/tmp" "azcopy.zip"
-unzip /tmp/azcopy.zip -d azcopy
-AZCOPY_EXTRACTED=$(echo azcopy/azcopy*)
-cp "$AZCOPY_EXTRACTED/azcopy" "/usr/local/bin/azcopy"
-chmod +x "/usr/local/bin/azcopy"
+    download_with_retries $AZCOPY_DOWNLOAD_URL "/tmp" "azcopy.zip"
+    unzip /tmp/azcopy.zip -d azcopy
+    AZCOPY_EXTRACTED=$(echo azcopy/azcopy*)
+    cp "$AZCOPY_EXTRACTED/azcopy" "/usr/local/bin/azcopy"
+    chmod +x "/usr/local/bin/azcopy"
 
-echo "Done, cleaning up"
-rm -rf azcopy*
+    echo "Done, cleaning up"
+    rm -rf azcopy*
+fi
 
 invoke_tests "Common" "AzCopy"
